Change instance type
Change uint8_t instance to uint32_t instance for whole project

+++ Libraries/drivers/K/inc/adc.h
@@ -139,10 +139,10 @@ typedef struct
 
 
 //!< API functions
-void ADC_CallbackInstall(uint8_t instance, ADC_CallBackType AppCBFun);
+void ADC_CallbackInstall(uint32_t instance, ADC_CallBackType AppCBFun);
 void ADC_Init(ADC_InitTypeDef* ADC_InitStruct);
 uint8_t ADC_QuickInit(uint32_t ADCxMAP, ADC_ResolutionMode_Type resolutionMode);
-void ADC_ITDMAConfig(uint8_t instance, uint32_t mux, ADC_ITDMAConfig_Type config);
+void ADC_ITDMAConfig(uint32_t instance, uint32_t mux, ADC_ITDMAConfig_Type config);
 int32_t ADC_QuickReadValue(uint32_t ADCxMAP);
 int32_t ADC_ReadValue(uint32_t instance, uint32_t mux);
 void ADC_StartConversion(uint32_t instance, uint32_t chl, uint32_t mux);
